Create my_last_decade_2014_to_2023
As we approach the end of 2023, I’m pausing to reflect on the remarkable journey of the past decade. Looking at the graph, it tells the story of my priorities in academics, Islamic pursuits, and extracurricular activities. Each line showcases my personal rating on a subjective 10-point scale, capturing the time and energy I invested in each facet of my life.